Move mock node to src
